requirements-dev.txt
-r requirements.txt
pytest>=8.3,<9.0
pytest-django>=4.7,<4.8
